+++ priv/static/live_monaco_editor.esm.js
@@ -495,7 +495,7 @@ var loader = {
 };
 var loader_default = loader;
 
-// js/live_monaco_editor/editor/themes.js
+// js/live_monaco_editor/editor/themes/livebook.js
 var colors = {
   background: "#282c34",
   default: "#c4cad6",
@@ -564,6 +564,90 @@ var theme = {
     "editorBracketMatch.background": "#3e4451"
   }
 };
+var livebook_default = theme;
+
+// js/live_monaco_editor/editor/themes/tokyonight.js
+var tokyonight_default = {
+  base: "vs-dark",
+  inherit: true,
+  rules: [
+    // Basic tokens
+    { token: "", foreground: "c0caf5" },
+    // Default text
+    { token: "comment", foreground: "565f89", fontStyle: "italic" },
+    { token: "comment.doc", foreground: "e0af68" },
+    // Documentation comments
+    // Constants
+    { token: "constant", foreground: "ff9e64" },
+    { token: "constant.language", foreground: "2ac3de" },
+    { token: "constant.numeric", foreground: "ff9e64" },
+    { token: "constant.character", foreground: "9ece6a" },
+    { token: "constant.character.escape", foreground: "bb9af7" },
+    // Strings
+    { token: "string", foreground: "9ece6a" },
+    { token: "string.regexp", foreground: "b4f9f8" },
+    { token: "string.special", foreground: "2ac3de" },
+    // Variables
+    { token: "variable", foreground: "c0caf5" },
+    { token: "variable.predefined", foreground: "f7768e" },
+    { token: "variable.parameter", foreground: "e0af68", fontStyle: "italic" },
+    { token: "member", foreground: "73daca" },
+    // Keywords
+    { token: "keyword", foreground: "9d7cd8", fontStyle: "italic" },
+    { token: "keyword.control", foreground: "bb9af7" },
+    { token: "keyword.operator", foreground: "bb9af7" },
+    { token: "keyword.import", foreground: "7dcfff" },
+    { token: "keyword.return", foreground: "9d7cd8", fontStyle: "italic" },
+    // Functions
+    { token: "function", foreground: "7aa2f7", fontStyle: "italic" },
+    { token: "function.builtin", foreground: "2ac3de" },
+    { token: "function.macro", foreground: "7dcfff" },
+    // Types
+    { token: "type", foreground: "2ac3de" },
+    { token: "type.builtin", foreground: "2ac3de" },
+    { token: "type.enum.variant", foreground: "ff9e64" },
+    { token: "namespace", foreground: "7dcfff" },
+    // Markup
+    { token: "markup.heading", foreground: "7aa2f7", fontStyle: "bold" },
+    { token: "markup.bold", fontStyle: "bold" },
+    { token: "markup.italic", fontStyle: "italic" },
+    { token: "markup.strikethrough", fontStyle: "strikethrough" },
+    { token: "markup.inserted", foreground: "9ece6a" },
+    { token: "markup.deleted", foreground: "f7768e" },
+    { token: "markup.changed", foreground: "6183bb" },
+    // Operators and punctuation
+    { token: "operator", foreground: "89ddff" },
+    { token: "punctuation", foreground: "89ddff" }
+  ],
+  colors: {
+    // Editor UI colors
+    "editor.background": "#1a1b26",
+    "editor.foreground": "#c0caf5",
+    "editor.lineHighlightBackground": "#292e42",
+    "editor.selectionBackground": "#283457",
+    "editor.selectionHighlightBackground": "#343a55",
+    "editorCursor.foreground": "#c0caf5",
+    "editorWhitespace.foreground": "#3b4261",
+    "editorLineNumber.foreground": "#3b4261",
+    "editorLineNumber.activeForeground": "#737aa2",
+    "editor.findMatchBackground": "#ff9e6455",
+    "editor.findMatchHighlightBackground": "#ff9e6422",
+    // UI elements
+    "editorWidget.background": "#16161e",
+    "editorWidget.border": "#15161e",
+    "editorSuggestWidget.background": "#16161e",
+    "editorSuggestWidget.border": "#15161e",
+    "editorSuggestWidget.selectedBackground": "#343a55",
+    "editorHoverWidget.background": "#16161e",
+    "editorHoverWidget.border": "#15161e",
+    "tab.activeBackground": "#1a1b26",
+    "tab.inactiveBackground": "#16161e",
+    "tab.activeForeground": "#c0caf5",
+    "tab.inactiveForeground": "#565f89",
+    "diffEditor.insertedTextBackground": "#449dab22",
+    "diffEditor.removedTextBackground": "#914c5422"
+  }
+};
 
 // js/live_monaco_editor/editor/code_editor.js
 var CodeEditor = class {
@@ -571,9 +655,11 @@ var CodeEditor = class {
     this.el = el;
     this.path = path;
     this.value = value;
-    this.opts = opts;
+    this.opts = opts || {};
+    this.theme = this.opts.theme || "tokyonight";
     this.standalone_code_editor = null;
     this._onMount = [];
+    this._monaco = null;
   }
   isMounted() {
     return !!this.standalone_code_editor;
@@ -596,13 +682,51 @@ var CodeEditor = class {
       this.standalone_code_editor.dispose();
     }
   }
+  /**
+   * Changes the editor theme
+   * @param {string} theme - The theme name ("default", "tokyonight")
+   * @returns {boolean} - Whether the theme was successfully changed
+   */
+  setTheme(theme2) {
+    this.theme = theme2;
+    if (this.isMounted() && this._monaco) {
+      try {
+        this._monaco.editor.setTheme(theme2);
+        return true;
+      } catch (error) {
+        console.error("Failed to set theme:", error);
+        return false;
+      }
+    }
+    return false;
+  }
+  /**
+   * Returns the available themes
+   * @returns {string[]} - Array of available theme names
+   */
+  getAvailableThemes() {
+    return ["default", "tokyonight"];
+  }
+  /**
+   * Returns the current theme
+   * @returns {string} - Current theme name
+   */
+  getCurrentTheme() {
+    return this.theme;
+  }
   _mountEditor() {
     this.opts.value = this.value;
     loader_default.config({
       paths: { vs: "https://cdn.jsdelivr.net/npm/monaco-editor@0.52.2/min/vs" }
     });
     loader_default.init().then((monaco) => {
-      monaco.editor.defineTheme("default", theme);
+      this._monaco = monaco;
+      monaco.editor.defineTheme("default", livebook_default);
+      monaco.editor.defineTheme("tokyonight", tokyonight_default);
+      monaco.editor.setTheme(this.theme);
+      console.log(monaco.editor);
+      console.log(tokyonight_default);
+      console.log(this.opts);
       let modelUri = monaco.Uri.parse(this.path);
       let language = this.opts.language;
       let model = monaco.editor.createModel(this.value, language, modelUri);